# jyv_tool/src/main/resources/yml/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Alquileres y Entidades
  description: API completa para la gestión de alquileres, devoluciones, entidades, herramientas, inventario, multas, pagos, reservas y usuarios.
  version: 1.0.0
servers:
  - url: http://localhost:8080/Api
    description: Servidor de desarrollo local
...
paths:
...
  /rol:
    get:
      summary: Obtener todos los roles
      description: Recupera una lista de todos los roles disponibles en el sistema.
      tags:
        - Entidades Generales
      responses:
        '200':
          description: Lista de roles exitosa.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rol'
              examples:
                rolesArray:
                  value:
                    - id: 1
                      nombre: "ADMIN"
                    - id: 2
                      nombre: "USUARIO"
        '500':
          description: Error interno del servidor.
...
  /herramienta:
    get:
      summary: Obtener todas las herramientas o filtrar por búsqueda/categoría
      description: |
        Recupera una lista de todas las herramientas. Permite filtrar por un término de búsqueda (`search`)
        o por categoría (`category`). Si ambos están presentes, se prioriza la búsqueda por término.
      tags:
        - Herramientas
      parameters:
        - name: search
          in: query
          description: Término de búsqueda para herramientas.
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Categoría de las herramientas (ej. "Electricidad", "Fontanería"). Usa "Todas" para ignorar el filtro.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista de herramientas exitosa.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Herramienta'
              examples:
                herramientasArray:
                  value:
                    - id: 1
                      nombre: "Taladro percutor"
                    - id: 2
                      nombre: "Sierra caladora"
        '500':
          description: Error interno del servidor.
    post:
      summary: Crear una nueva herramienta
      description: Registra una nueva herramienta en el sistema.
      tags:
        - Herramientas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HerramientaRequest'
            example:
              nombre: "Martillo demoledor"
              descripcion: "Martillo eléctrico para trabajos pesados."
              categoria: "Construcción"
              precioAlquilerDia: 45.00
      responses:
        '200':
          description: Herramienta creada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseHerramientas'
              example:
                success: true
                message: "Herramienta creada/actualizada."
                data:
                  id: 1
                  nombre: "Martillo demoledor"
                  categoria: "Construcción"
        '400':
          description: Solicitud inválida.
        '500':
          description: Error interno del servidor.
  /herramienta/{id}:
    get:
      summary: Obtener una herramienta por ID
      description: Recupera una herramienta específica por su ID.
      tags:
        - Herramientas
      parameters:
        - name: id
          in: path
          description: ID de la herramienta a obtener.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Herramienta encontrada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Herramienta'
              example:
                id: 1
                nombre: "Taladro percutor"
                descripcion: "Taladro eléctrico de alta potencia."
                categoria: "Herramientas Eléctricas"
                precioAlquilerDia: 25.00
                disponible: true
        '404':
          description: Herramienta no encontrada.
        '500':
          description: Error interno del servidor.
    delete:
      summary: Eliminar una herramienta por ID
      description: Elimina una herramienta específica del sistema.
      tags:
        - Herramientas
      parameters:
        - name: id
          in: path
          description: ID de la herramienta a eliminar.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Herramienta eliminada exitosamente.
        '404':
          description: Herramienta no encontrada.
        '500':
          description: Error interno del servidor.
    patch:
      summary: Actualizar una herramienta existente
      description: Actualiza parcialmente la información de una herramienta específica.
      tags:
        - Herramientas
      parameters:
        - name: id
          in: path
          description: ID de la herramienta a actualizar.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HerramientaRequest'
            example:
              precioAlquilerDia: 30.00
      responses:
        '200':
          description: Herramienta actualizada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseHerramientas'
              example:
                success: true
                message: "Herramienta actualizada exitosamente."
                data:
                  id: 1
                  nombre: "Taladro percutor"
                  categoria: "Herramientas Eléctricas"
                  precioAlquilerDia: 30.00
        '400':
          description: Solicitud inválida.
        '404':
          description: Herramienta no encontrada.
        '500':
          description: Error interno del servidor.

  # --- InventarioController ---
...
components:
  schemas:
...
    Rol:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único del rol.
          example: 1
        nombre:
          type: string
          description: Nombre del rol (ej. ADMIN, USUARIO).
          example: "ADMIN"
      example:
        id: 1
        nombre: "ADMIN"
...
    Herramienta:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único de la herramienta.
          example: 1
        nombre:
          type: string
          description: Nombre de la herramienta.
          example: "Taladro percutor"
        descripcion:
          type: string
          description: Descripción de la herramienta.
          example: "Taladro eléctrico de alta potencia."
        categoria:
          type: string
          description: Categoría de la herramienta.
          example: "Herramientas Eléctricas"
        precioAlquilerDia:
          type: number
          format: float
          description: Precio de alquiler por día.
          example: 25.00
        disponible:
          type: boolean
          description: Indica si la herramienta está disponible.
          example: true
      example:
        id: 1
        nombre: "Taladro percutor"
        descripcion: "Taladro eléctrico de alta potencia."
        categoria: "Herramientas Eléctricas"
        precioAlquilerDia: 25.00
        disponible: true
...
    HerramientaRequest:
      type: object
      required:
        - nombre
        - categoria
        - precioAlquilerDia
      properties:
        nombre:
          type: string
          description: Nombre de la herramienta.
          example: "Martillo demoledor"
        descripcion:
          type: string
          description: Descripción de la herramienta.
          example: "Martillo eléctrico para trabajos pesados."
        categoria:
          type: string
          description: Categoría a la que pertenece la herramienta.
          example: "Construcción"
        precioAlquilerDia:
          type: number
          format: float
          description: Precio de alquiler por día.
          example: 45.00
...
    ResponseHerramientas:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Herramienta creada/actualizada."
        data:
          $ref: '#/components/schemas/Herramienta'
...